Updated paths to standard naming.
