+++ src/stochastic/classes/RSA1D/nn_exclusion.py
@@ -1,25 +1,23 @@
-""" File that contains the random sequential adsorption of dimers.
+""" File that contains the random sequential adsorption with nearest neighbors
+    exclusion.
 """
 
 # ------------------------------------------------------------------------------
 # Imports.
 # ------------------------------------------------------------------------------
 
-# Imports: General.
-import numpy
-
 # Imports: User-defined.
-from stochastic.Interfaces.RSA_2D_interface import RSA2D
-from stochastic.Utilities.RSA_parameters import RSA2DParameters
+from stochastic.interfaces.RSA_1D_interface import RSA1D
+from stochastic.utilities.RSA_parameters import RSA1DParameters
 
 # ------------------------------------------------------------------------------
 # Classes.
 # ------------------------------------------------------------------------------
 
 
-class Dimers(RSA2D):
-    """ Class to simulate random sequential adsorption of dimers for a
-        two-dimensional lattice.
+class NNExclusion(RSA1D):
+    """ Class to simulate random sequential adsorption with nearest neighbor
+        exclusion for a one-dimensional lattice.
 
         Inherited parameters:
 
@@ -60,7 +58,7 @@ class Dimers(RSA2D):
     # --------------------------------------------------------------------------
     # Get Methods.
     # --------------------------------------------------------------------------
-    
+
     def get_preheader(self) -> str:
         """ Returns the pre-header, i.e., the string that contains the
             simulation information.
@@ -69,8 +67,8 @@ class Dimers(RSA2D):
         """
 
         preheader = ",".join([
-            f"RSA-2D Dimer", f"seed={self.seed}", f"repetitions(n)={self.repetitions}",
-            f"maximum time={self.maximum_time}", f"dimensions={self.dimensions}", f"periodic={self.periodic}"
+            f"RSA Nearest Neighbor Exclusion", f"seed={self.seed}", f"repetitions(n)={self.repetitions}",
+            f"maximum time={self.maximum_time}", f"length={self.length}", f"periodic={self.periodic}"
         ])
 
         return preheader
@@ -84,53 +82,41 @@ class Dimers(RSA2D):
             in the lattice onto which to adsorb.
         """
 
-        generator = self.random_generator.choice
-        site_0 = [0, 0]
-        site_1 = generator(numpy.array([[1, 0], [0, 1]], dtype=int))
-
-        generator = self.random_generator.integers
-        if numpy.array_equal(site_1, numpy.array([1, 0], dtype=int)):
-            site_0[0] = generator(0, self.dimensions[0]) if self.periodic[0] else generator(0, self.dimensions[0] - 1)
-            site_0[1] = generator(0, self.dimensions[1])
-        else:
-            site_0[0] = generator(0, self.dimensions[0])
-            site_0[1] = generator(0, self.dimensions[1]) if self.periodic[1] else generator(0, self.dimensions[1] - 1)
-
-        site_1 = tuple(map(int, site_0 + site_1))
-        site_1 = self.normalize_site(site_1)
-
+        site = self.random_generator.integers(0, self.length)
         self.attempts += 1
-        site_0 = tuple(site_0)
-        if self.validate_adsorb(site_0, site_1):
-            self.lattice[site_0[0]][site_0[1]] = RSA2D.OCCUPIED
-            self.lattice[site_1[0]][site_1[1]] = RSA2D.OCCUPIED
+        if self.validate_adsorb(site):
+            self.lattice[site] = RSA1D.OCCUPIED
             self.attempts_successful += 1
 
     # --------------------------------------------------------------------------
     # Validate Methods.
     # --------------------------------------------------------------------------
 
-    def validate_adsorb(self, site_0: tuple, site_1: tuple) -> bool:
+    def validate_adsorb(self, site: int) -> bool:
         """ Determines if the given site can adsorb a particle.
 
-            :param site_0: Number of the zeroth site where the dimer is going to
-             be adsorbed.
-
-             :param site_1: Number of the first site where the dimer is going to
-             be adsorbed.
+            :param site: The site to be examined. Must be an integer number.
 
             :return: If the site is empty and its inmediate neighbors are empty.
         """
 
-        if not self.lattice[site_0[0]][site_0[1]] == RSA2D.EMPTY:
+        site_ = self.normalize_site(site)
+        if not self.lattice[site_] == RSA1D.EMPTY:
             return False
 
-        examined = {site_0}
-        if self.validate_in_lattice(site_1) and site_1 not in examined:
-            if self.lattice[site_1[0]][site_1[1]] == RSA2D.EMPTY:
-                return True
+        examined = {site_}
+        if ((site - 1) < 0 and self.periodic) or site > 0:
+            site_ = self.normalize_site(site - 1)
+            if (site_ not in examined) and not self.lattice[site_] == RSA1D.EMPTY:
+                return False
 
-        return False
+        examined.add(site_)
+        if ((site + 1) >= self.length and self.periodic) or (site + 1) < self.length:
+            site_ = self.normalize_site(site + 1)
+            if (site_ not in examined) and not self.lattice[site_] == RSA1D.EMPTY:
+                return False
+
+        return True
 
     # $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
     # Constructor and Dunder Methods.
@@ -140,10 +126,10 @@ class Dimers(RSA2D):
     # Constructor.
     # --------------------------------------------------------------------------
 
-    def __init__(self, parameters: RSA2DParameters):
+    def __init__(self, parameters: RSA1DParameters):
         """ Initializes the simulation parameters.
 
             :param parameters: A dataclass that contains the adjustable
              parameters of the simulation.
         """
-        super(Dimers, self).__init__(parameters)
+        super(NNExclusion, self).__init__(parameters)
